build: Remove bzip2 tarballs from distcheck
When adding xz as third compression format in May 2014 I mentioned that I'm
going to remove bzip2 in favor of xz in 1.12.
I heard no complains but I left it in for 1.12. With 1.13 starting its time
to remove it. From now on it will be only gz and xz compressed tarballs.